Simplify rollout by not requiring user to specify fact we can gather
Before, the user had to specify the SSH pubkey manually, which would be
used by the backup server to connect to backup clients.
We can gather this during runtime, as we have to connect to the backup
server in the same way later during the playbook anyways.
This makes it easier hopefully.