Set strict defaults, as they are "defaulted" in the other roles.
There was an issue where this _defaults_ file would be overridden with
a hosts custom values, if run against this single host with
'ansible-playbook --limit=host.example.tld', or even if the first host
of the running batch has custom values.

This is of course not what this file intends to do.